refactor(edge/groups): break form to components
