+++ app/react/edge/edge-groups/components/EdgeGroupForm/group-type-options.tsx
