+++ app/react/edge/edge-groups/components/EdgeGroupForm/types.tsx
@@ -1,0 +1,10 @@
+import { EnvironmentId } from '@/react/portainer/environments/types';
+import { TagId } from '@/portainer/tags/types';
+
+export interface FormValues {
+  name: string;
+  dynamic: boolean;
+  environmentIds: EnvironmentId[];
+  partialMatch: boolean;
+  tagIds: TagId[];
+}
